Fix bug leaving empty file
Fix a bug that creates the empty `dst_filepath' file, but without
writing to it. This happens when calling fopen on the `src_filepath',
but an error occurs (eg. file is not found). The check for this takes
places _after_ fopen is called on `dst_filepath', leaving the file
created, but without any content to write to it.